requirements.txt
streamlit
qrcode
Pillow
pyzbar
numpy